% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCIItocsv.R
\name{ASCIItocsv}
\alias{ASCIItocsv}
\title{Formatting K-NET/KiK-net ASCII to csv file}
\usage{
ASCIItocsv(flist, exdir, unified = TRUE)
}
\arguments{
\item{flist}{list of pathnames of K-NET/KiK-net ASCII format file}

\item{exdir}{directory for output}
}
\description{
Formatting K-NET/KiK-net ASCII to csv file
}
