% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractASCII.R
\name{extractASCII}
\alias{extractASCII}
\title{Extracting K-NET/KiK-net ASCII format files from compressed folder}
\usage{
extractASCII(tarfile, exdir)
}
\arguments{
\item{tarfile}{compressed .tar.gz folder}

\item{exdir}{directory for output}
}
\description{
Extracting K-NET/KiK-net ASCII format files from compressed folder
}
