% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp_dist.R
\name{hyp_dist}
\alias{hyp_dist}
\title{Calculating hypocentral distance between source and station.}
\usage{
hyp_dist(depth, epi_N, epi_E, sta_N, sta_E, r = 6371)
}
\arguments{
\item{depth}{the depth of hypocenter.}

\item{epi_N}{latitude of epicenter.}

\item{epi_E}{longitude of epicenter.}

\item{sta_N}{latitude of station.}

\item{sta_E}{longitude of station.}

\item{r}{radius of the earth (default = 6371).}
}
\description{
Calculating hypocentral distance between source and station.
}
