% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnetDataListing.R
\name{KnetDataListing}
\alias{KnetDataListing}
\title{Listing the properties of K-NET/KiK-net ASCII format files}
\usage{
KnetDataListing(flist)
}
\arguments{
\item{flist}{list of pathnames of K-NET/KiK-net ASCII format file}
}
\description{
Listing the properties of K-NET/KiK-net ASCII format files
}
