% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnetDataSearch.R
\name{KnetDataSearch}
\alias{KnetDataSearch}
\title{Searching K-NET/KiK-net ASCII format files by given conditions}
\usage{
KnetDataSearch(
  flist,
  olat = c(0, 90),
  olong = c(0, 180),
  depth = c(0, 10^4),
  mag = c(0, 10),
  slat = c(0, 90),
  slong = c(0, 180),
  shgt = c(-10^8, 10^8),
  sfreq = c(100, 200),
  dur = c(0, 10^4),
  macc = c(0, 10^5),
  scode = c(),
  direc = c("N-S", "E-W", "U-D", 1:6),
  epid = c(0, 10^5),
  hypd = c(0, 10^5),
  otime = as.POSIXlt(c("1900-01-01 00:00:00 JST", "2100-01-01 00:00:00 JST")),
  rtime = as.POSIXlt(c("1900-01-01 00:00:00 JST", "2100-01-01 00:00:00 JST")),
  ctime = as.POSIXlt(c("1900-01-01 00:00:00 JST", "2100-01-01 00:00:00 JST"))
)
}
\arguments{
\item{flist}{list of pathnames of K-NET/KiK-net ASCII format file}
}
\description{
Searching K-NET/KiK-net ASCII format files by given conditions
}
