% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnetReader2.R
\name{KnetReader2}
\alias{KnetReader2}
\title{Reading K-NET/KiK-net ASCII format files}
\usage{
KnetReader2(fname)
}
\arguments{
\item{fname}{filename (or pathname) of K-NET/KiK-net ASCII format file}
}
\description{
Reading K-NET/KiK-net ASCII format files
}
